Refactoring duplicate code SonarCloud
