--- src/main/resources/lib/types/partTypes/statisticDescription.ts
@@ -1,10 +1,0 @@
-import { type AccordionData } from '/lib/types/partTypes/accordion'
-
-export interface StatisticDescriptionProps {
-  icon: string
-  ingress: string
-  label: string
-  lastUpdatedPhrase: string
-  lastUpdated: string
-  accordions: AccordionData[]
-}
